move bitbucket server logic to separate file and few code improvements
